start make fav fragm

+++ app/src/main/java/com/medvedev/financeapp/presentation/viewModel/FavouriteViewModel.java
@@ -33,7 +33,7 @@ public class FavouriteViewModel  extends AndroidViewModel {
     }
 
     public void addStock(String ticker,
-                         int lastCost,
+                         Double lastCost,
                          String companyName,
                          int cost){
